move cloud cli into subdirectory
