+++ src/prefect/cli/__init__.py
@@ -8,7 +8,8 @@ from prefect.cli.root import app
 import prefect.cli.agent
 import prefect.cli.artifact
 import prefect.cli.block
-import prefect.cli.cloud
+import prefect.cli.cloud.cloud
+import prefect.cli.cloud.webhook
 import prefect.cli.concurrency_limit
 import prefect.cli.config
 import prefect.cli.deploy
