+++ src/prefect/cli/cloud/webhook.py
@@ -3,12 +3,12 @@ Command line interface for working with webhooks
 """
 
 from prefect.cli._types import PrefectTyper
-from prefect.cli.root import app
+from prefect.cli.cloud.cloud import cloud_app
 
 webhook_app = PrefectTyper(
     name="webhook", help="Commands for starting and interacting with webhooks"
 )
-app.add_typer(webhook_app)
+cloud_app.add_typer(webhook_app)
 
 
 @webhook_app.command()
